# 🚀 Major Performance Optimizations & System Enhancements v2.1