--- a/logic/mode_1.py
+++ b/logic/mode_1.py
@@ -4,13 +4,13 @@
 # enriched versions with dynamic output length control.
 
 from typing import Optional, Dict, Union
-from utils.generator import GroqGenerator
+from utils.generator import get_generator
 from utils.validator import calculate_max_tokens
 
-class Mode1Logic:    
+class Mode1Logic:
     def __init__(self):
-        # Initialize the Groq LLM generator for text completion
-        self.generator = GroqGenerator()
+        # Use singleton generator for better performance
+        self.generator = get_generator()
     
     def get_system_prompt(self) -> str:
         return (